requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
distlib==0.3.9
Django==4.2.22
django-stubs==5.2.0
django-stubs-ext==5.2.0
djangorestframework==3.16.0
djangorestframework-stubs==3.16.0
filelock==3.18.0
identify==2.6.12
idna==3.10
mypy==1.15.0
mypy_extensions==1.1.0
nodeenv==1.9.1
pathspec==0.12.1
platformdirs==4.3.8
pre_commit==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
ruff==0.11.13
six==1.17.0
sqlparse==0.5.3
types-python-dateutil==2.9.0.20250516
types-PyYAML==6.0.12.20250516
types-requests==2.32.0.20250602
typing_extensions==4.14.0
urllib3==2.4.0
virtualenv==20.31.2
